updated code to working version #28
