+++ .github/workflows/tests.yml
@@ -39,9 +39,15 @@ jobs:
       - name: Cache hit
         run: echo '${{ steps.pysetup.outputs.cache-hit }}'
 
+      - name: FFTW dependencies
+        run: |
+          sudo apt update
+          sudo apt install libfftw3-dev -y
+
       # build environment with pip
       - name: Install ffpiv
         run: |
+          
           pip install --upgrade pip
           pip install .[test,extra]
